Add Elixir solution based on List.foldl/3
With assumption that everything is a List, we only need to fold it right!

As we also need to track additional data (section or lesson positions)
between the sections, we use a two elements tuple to pass these pieces of
information between the folds.

Though this little tuple-hack makes us to use `elem/2` before we can return
a true output result (as we have also to reverse accumulated lists).

The similar approach with tuple-based accumulator we use for the lessons
folding flow.