Add a common lisp solution using LOOP.
More than a bit more imperative than might be expected from LISP.  I can't
make legitimate claims that this is idiomatic, but it is the first that
came to mind on how I'd do this with lisp.  (I confess this came a bit more
naturally in elisp, but alas.)

This does have a bit of a feel that I was going for code golf, but I argue
against that idea.  This is the natural result of treating the data as a
nested set of plists.  If this was something that was to be more long
lived, I think it would naturally extend to some defstructs, but the
general loop would remain the same.