Update SML example to eliminate use of the cat operator (`@`)
An expression of the form `xs @ [...]` rebuilds a whole new list.  Doing
this on each recursion leads to O(n^2) behaviour.  The idiomatic
approach is to accumulate a reversed list with `foldl` using cons and
reverse it at the end.  Although `foldr` would allow a list to be
accumulated without being reversed, its use is discouraged on data
structures of unbounded size because it is not tail recursive.